+++ addons/website_profile/models/res_users.py
@@ -14,6 +14,15 @@ VALIDATION_KARMA_GAIN = 3
 class Users(models.Model):
     _inherit = 'res.users'
 
+    def __init__(self, pool, cr):
+        init_res = super(Users, self).__init__(pool, cr)
+        type(self).SELF_WRITEABLE_FIELDS = list(
+            set(
+                self.SELF_WRITEABLE_FIELDS +
+                ['country_id', 'city', 'website', 'website_description', 'website_published']))
+        type(self).SELF_READABLE_FIELDS = type(self).SELF_READABLE_FIELDS + ['karma']
+        return init_res
+
     @api.model
     def _generate_profile_token(self, user_id, email):
         """Return a token for email validation. This token is valid for the day
