+++ addons/website_profile/views/website_profile.xml
@@ -497,6 +497,7 @@
                     <img class="position-absolute" t-attf-src="/website_profile/static/src/img/rank_#{user['position']}.svg" alt="User rank" style="bottom: 0; right: -10px"/>
                 </div>
                 <h3 class="mt-2 mb-0" t-esc="user['name']"></h3>
+                <span class="badge badge-danger font-weight-normal px-2" t-if="not user['website_published']">Unpublished</span>
                 <strong class="text-muted" t-esc="user['rank']"/>
             </div>
             <div class="row mx-0 o_wprofile_top3_card_footer text-nowrap">
@@ -516,6 +517,9 @@
         <td class="align-middle w-md-75">
             <span class="font-weight-bold" t-esc="user['name']"/><br/>
             <span class="text-muted font-weight-bold" t-esc="user['rank']"></span>
+        </td>
+         <td t-if="not user['website_published']" class="align-middle font-weight-bold text-right text-nowrap">
+            <span class="badge badge-danger font-weight-normal px-2 py-1 m-1">Unpublished</span>
         </td>
         <td class="align-middle font-weight-bold text-right text-nowrap">
             <b t-esc="user['karma']"/> <span class="text-muted small font-weight-bold">XP</span>
